+++ flutter_navigation_generator/lib/src/code_builder/on_generate_route_builder.dart
@@ -2,6 +2,7 @@ import 'package:code_builder/code_builder.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter_navigation_generator/src/models/importable_type.dart';
 import 'package:flutter_navigation_generator/src/models/route_config.dart';
+import 'package:flutter_navigation_generator/src/models/route_field_config.dart';
 import 'package:flutter_navigation_generator/src/utils/case_utils.dart';
 import 'package:flutter_navigation_generator/src/utils/importable_type_string_converter.dart';
 import 'package:flutter_navigation_generator/src/utils/route_config_extension.dart';
@@ -14,12 +15,14 @@ class OnGenerateRouteBuilder {
   final ImportableType? unknownRoute;
   final List<ImportableType> defaultGuards;
   final Uri? targetFile;
+  final bool ignoreKeysByDefault;
 
   OnGenerateRouteBuilder({
     required this.routes,
     this.pageType,
     this.unknownRoute,
     this.targetFile,
+    this.ignoreKeysByDefault = true,
     this.defaultGuards = const [],
   });
 
@@ -36,17 +39,10 @@ class OnGenerateRouteBuilder {
     final constructor = route.constructorName == route.routeWidget.className || route.constructorName.isEmpty
         ? route.routeWidget.className
         : '${route.routeWidget.className}.${route.constructorName}';
-    final constructorCall = '$constructor(${route.parameters.asMap().map((_, parameterConfig) {
-          final parameter = parameterConfig.type;
-          final nullableSuffix = parameter.isNullable ? '?' : '';
-          final convertFromString = ImportableTypeStringConverter.convertFromString(parameter, 'arguments[\'${parameter.argumentName}\']');
-          return MapEntry(
-            parameter.argumentName,
-            (parameter.className == 'Key' || parameter.className == 'String')
-                ? "arguments['${parameter.argumentName}'] as ${parameter.className}$nullableSuffix"
-                : "arguments['${parameter.argumentName}'] is String ? $convertFromString : arguments['${parameter.argumentName}'] as ${parameter.className}${parameter.typeArguments.isNotEmpty ? '<${parameter.typeArguments.map((e) => e.className).join(',')}>' : ''}$nullableSuffix",
-          );
-        }).entries.map((e) => '${e.key}: ${e.value},').join('')})';
+    final constructorCall = '$constructor(${route.parameters.asMap().map((_, parameterConfig) => MapEntry(
+          parameterConfig.type.argumentName,
+          _getParameterValue(parameterConfig),
+        )).entries.where((e) => e.value != null).map((e) => '${e.key}: ${e.value},').join('')})';
 
     var guardsCode = '';
     if ((route.guards ?? defaultGuards).isNotEmpty) {
@@ -63,6 +59,29 @@ class OnGenerateRouteBuilder {
       }
     }
     return '${guardsCode}return ${_withPageType(route, constructorCall)};';
+  }
+
+  String? _getParameterValue(RouteFieldConfig parameterConfig) {
+    if (parameterConfig.ignoreWithKeyCheck(ignoreKeysByDefault) == true) return null;
+    final parameter = parameterConfig.type;
+    final nullableSuffix = parameter.isNullable ? '?' : '';
+    final defaultSuffix = parameterConfig.defaultValue != null ? '${parameter.isNullable ? '' : '?'} ?? ${parameterConfig.defaultValue}' : '';
+    final convertFromString = ImportableTypeStringConverter.convertFromString(parameter, 'queryParameters[\'${parameterConfig.queryName}\']!');
+    if (parameter.className == 'Key') return "arguments['${parameter.argumentName}'] as ${parameter.className}$nullableSuffix$defaultSuffix";
+    // Small notation in case of string/dynamic (using ??)
+    if (parameter.className == 'String' || parameter.className == 'dynamic') {
+      return "queryParameters['${parameterConfig.queryName}'] ?? arguments['${parameter.argumentName}'] as ${parameter.className}$nullableSuffix$defaultSuffix";
+    }
+
+    final typeArguments = parameter.typeArguments.isNotEmpty ? '<${parameter.typeArguments.map((e) => e.className).join(',')}>' : '';
+    final String fromArguments;
+    if (parameterConfig.defaultValue != null) {
+      fromArguments = "arguments['${parameter.argumentName}'] as ${parameter.className}$typeArguments? ?? ${parameterConfig.defaultValue}";
+    } else {
+      fromArguments = "arguments['${parameter.argumentName}'] as ${parameter.className}$typeArguments$nullableSuffix";
+    }
+    if (parameterConfig.addToJson == false) return fromArguments;
+    return "queryParameters['${parameterConfig.queryName}'] != null ? $convertFromString : $fromArguments";
   }
 
   Method generate() {
@@ -84,60 +103,62 @@ class OnGenerateRouteBuilder {
               ..type = const Reference('RouteSettings'),
           ),
         )
-        ..body = Block.of([
-          if (pageRoutes.isNotEmpty) ...[
-            const Code('''final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};
-    final settingsUri = Uri.parse(settings.name ?? '');
-    settingsUri.queryParameters.forEach((key, value) {
-      arguments[key] ??= value;
-    });'''),
-            Code(
-                'switch (settingsUri.path) {${pageRoutes.where((route) => !route.routeNameContainsParameters).map((route) => 'case RouteNames.${route.asRouteName}: ${_generateRoute(route)}').join('')}}'),
-          ],
-          if (pageRoutes.any((element) => element.routeNameContainsParameters)) ...[
-            const Code('final pathSegments = settingsUri.pathSegments;'),
-            ...pageRoutes
-                .where((pageRoute) => pageRoute.routeNameContainsParameters)
-                .groupListsBy((pageRoute) => pageRoute.routeName.pathSegments.length)
-                .entries
-                .sorted((a, b) => -a.key.compareTo(b.key))
-                .map((group) {
-              final pathSegments = group.key;
-              var code = 'if (pathSegments.length == $pathSegments) {';
-              final pageRoutesMap = group.value.asMap().map((key, value) => MapEntry(value, value.routeName.parametersFromRouteName.length));
-              final pageRoutes = pageRoutesMap.entries.sorted((a, b) => a.value.compareTo(b.value)).map((e) => e.key).toList();
-              for (final pageRoute in pageRoutes) {
-                final pathSegments = pageRoute.routeName.pathSegments;
-                final hasRigidSegments = pathSegments.any((element) => !element.startsWith(':'));
-                if (hasRigidSegments) {
-                  code += 'if (';
-                  code += pathSegments
-                      .asMap()
-                      .entries
-                      .where((pathSegment) => !pathSegment.value.startsWith(':'))
-                      .map((pathSegment) => 'pathSegments[${pathSegment.key}] == \'${pathSegment.value}\'')
-                      .join(' && ');
-                  code += ') {';
-                }
-                code += pathSegments
-                    .asMap()
-                    .entries
-                    .where((pathSegment) => pathSegment.value.startsWith(':'))
-                    .map((pathSegment) => 'arguments[\'${pathSegments[pathSegment.key].substring(1)}\'] = pathSegments[${pathSegment.key}];')
-                    .join('\n');
-                code += _generateRoute(pageRoute);
-                if (hasRigidSegments) code += '}';
-              }
-
-              return Code('$code}');
-            }),
-          ],
-          if (unknownRoute != null) ...[
-            Code('return ${_withPageType(null, '${typeRefer(unknownRoute!).symbol!}()')};'),
-          ] else ...[
-            const Code('return null;'),
-          ],
-        ]),
+        ..body = _generateBody(pageRoutes),
     );
+  }
+
+  Block _generateBody(List<RouteConfig> pageRoutes) {
+    return Block.of([
+      if (pageRoutes.isNotEmpty) ...[
+        const Code('''final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};
+    final settingsUri = Uri.parse(settings.name ?? '');
+    final queryParameters = Map.from(settingsUri.queryParameters);'''),
+        Code(
+            'switch (settingsUri.path) {${pageRoutes.where((route) => !route.routeNameContainsParameters).map((route) => 'case RouteNames.${route.asRouteName}: ${_generateRoute(route)}').join('')}}'),
+      ],
+      if (pageRoutes.any((element) => element.routeNameContainsParameters)) ...[
+        const Code('final pathSegments = settingsUri.pathSegments;'),
+        ...pageRoutes
+            .where((pageRoute) => pageRoute.routeNameContainsParameters)
+            .groupListsBy((pageRoute) => pageRoute.routeName.pathSegments.length)
+            .entries
+            .sorted((a, b) => -a.key.compareTo(b.key))
+            .map((group) {
+          final pathSegments = group.key;
+          var code = 'if (pathSegments.length == $pathSegments) {';
+          final pageRoutesMap = group.value.asMap().map((key, value) => MapEntry(value, value.routeName.parametersFromRouteName.length));
+          final pageRoutes = pageRoutesMap.entries.sorted((a, b) => a.value.compareTo(b.value)).map((e) => e.key).toList();
+          for (final pageRoute in pageRoutes) {
+            final pathSegments = pageRoute.routeName.pathSegments;
+            final hasRigidSegments = pathSegments.any((element) => !element.startsWith(':'));
+            if (hasRigidSegments) {
+              code += 'if (';
+              code += pathSegments
+                  .asMap()
+                  .entries
+                  .where((pathSegment) => !pathSegment.value.startsWith(':'))
+                  .map((pathSegment) => 'pathSegments[${pathSegment.key}] == \'${pathSegment.value}\'')
+                  .join(' && ');
+              code += ') {';
+            }
+            code += pathSegments
+                .asMap()
+                .entries
+                .where((pathSegment) => pathSegment.value.startsWith(':'))
+                .map((pathSegment) => 'queryParameters[\'${pathSegments[pathSegment.key].substring(1)}\'] = pathSegments[${pathSegment.key}];')
+                .join('\n');
+            code += _generateRoute(pageRoute);
+            if (hasRigidSegments) code += '}';
+          }
+
+          return Code('$code}');
+        }),
+      ],
+      if (unknownRoute != null) ...[
+        Code('return ${_withPageType(null, '${typeRefer(unknownRoute!).symbol!}()')};'),
+      ] else ...[
+        const Code('return null;'),
+      ],
+    ]);
   }
 }