+++ flutter_navigation_generator/lib/src/resolvers/route_field_resolver.dart
@@ -13,7 +13,7 @@ class RouteFieldResolver {
 
   RouteFieldResolver(List<LibraryElement> libs) : _typeResolver = ImportableTypeResolverImpl(libs);
 
-  List<RouteFieldConfig> resolveConstructorFieldsMethod(ExecutableElement constructor) {
+  List<_RawRouteFieldConfig> _resolveConstructorFields(ExecutableElement constructor) {
     final constructorParameters = constructor.parameters;
 
     return constructorParameters.map((parameter) {
@@ -24,57 +24,73 @@ class RouteFieldResolver {
       );
       final annotation = _flutterRouteFieldAnnotationChecker.annotationsOf(parameter, throwOnUnresolved: false).map(ConstantReader.new);
       if (annotation.isEmpty) {
-        return RouteFieldConfig(
+        return _RawRouteFieldConfig(
           type: field,
           defaultValue: parameter.defaultValueCode,
-          queryName: field.name!,
-          ignore: false,
-          addToJson: true,
         );
       }
-      return getConfigFromField(
-        annotation,
-        null,
-        field,
+      final flutterRouteField = annotation.firstOrNull;
+      final queryName = flutterRouteField?.peek('queryName')?.stringValue;
+      final ignore = flutterRouteField?.peek('ignore')?.boolValue;
+      final addToJson = flutterRouteField?.peek('addToJson')?.boolValue;
+
+      return _RawRouteFieldConfig(
+        type: field,
+        defaultValue: parameter.defaultValueCode,
+        queryName: queryName,
+        ignore: ignore,
+        addToJson: addToJson,
       );
     }).toList();
   }
 
-  List<RouteFieldConfig> resolveFieldsMethod(ClassElement classElement, List<RouteFieldConfig> fields) {
+  List<RouteFieldConfig> resolveFieldsMethod(
+    ExecutableElement constructor,
+    ClassElement classElement,
+  ) {
     final fieldsWithAnnotation =
         classElement.fields.asMap().map((key, value) => MapEntry(value.displayName, _flutterRouteFieldAnnotationChecker.annotationsOf(value).map(ConstantReader.new)));
 
-    return fieldsWithAnnotation.entries.map((entry) {
-      final field = entry.key;
-      final fieldData = fields.firstWhere((element) => element.queryName == field);
-      final annotations = entry.value;
-
-      return getConfigFromField(
-        annotations,
-        fieldData,
-        null,
-      );
-    }).toList();
+    final constructorParameters = _resolveConstructorFields(constructor);
+    return constructorParameters
+        .map((parameter) => _getConfigFromField(
+              fieldsWithAnnotation[parameter.type.argumentName],
+              parameter,
+            ))
+        .toList();
   }
 
-  RouteFieldConfig getConfigFromField(
-    Iterable<ConstantReader> annotations,
-    RouteFieldConfig? fieldData,
-    ImportableType? type,
+  RouteFieldConfig _getConfigFromField(
+    Iterable<ConstantReader>? annotations,
+    _RawRouteFieldConfig parameterData,
   ) {
-    assert(fieldData != null || type != null);
-    final annotation = annotations.firstOrNull;
-    final flutterRouteField = annotation?.read('flutterRouteField');
+    final flutterRouteField = annotations?.firstOrNull;
     final queryName = flutterRouteField?.peek('queryName')?.stringValue;
     final ignore = flutterRouteField?.peek('ignore')?.boolValue;
     final addToJson = flutterRouteField?.peek('addToJson')?.boolValue;
 
     return RouteFieldConfig(
-      type: fieldData?.type ?? type!,
-      defaultValue: fieldData?.defaultValue,
-      queryName: queryName ?? fieldData?.queryName ?? type!.name!,
-      ignore: ignore ?? fieldData?.ignore ?? (type!.className == 'Key' && type.name == 'key'),
-      addToJson: addToJson ?? fieldData?.addToJson ?? true,
+      type: parameterData.type,
+      defaultValue: parameterData.defaultValue,
+      queryName: parameterData.queryName ?? queryName ?? parameterData.type.argumentName,
+      ignore: parameterData.ignore ?? ignore,
+      addToJson: parameterData.addToJson ?? addToJson ?? true,
     );
   }
+}
+
+class _RawRouteFieldConfig {
+  final ImportableType type;
+  final String? defaultValue;
+  final bool? ignore;
+  final bool? addToJson;
+  final String? queryName;
+
+  _RawRouteFieldConfig({
+    required this.type,
+    required this.defaultValue,
+    this.ignore,
+    this.addToJson,
+    this.queryName,
+  });
 }